1293224: char font
just preparing for a bigger change...
introduced a global font (just one), it is lower case per default.

I intend to modify the function for registering the default xpm to something
accepting the 64x64 bitmap (application dependent) and insert it into the
global pixmap.  an other function could be used to register the alfabet
pixmap and correspondently setting the variables indicating the position of
each character.  we would only support monospace alfabets.

the 'printing' functions would also be simpler (less arguments), since it
would be the library knowing which is the alfabet in use.