1587104: error initializing font 3x5
it really was a library problem...  it was caused inside getFont by the way
I was choosing an alternative to a clashing character code.  the correct way
is by choosing outside both the already defined characters and the
characters that the font uses.  we were creating a cascade in clashes, and
were not able to handle it correctly.  avoiding the problem is the best
cure.