1714519 - object oriented library
added new example application, on the way to the object oriented library.

the idea is to develop this program within one single source, but clearly
separating the general from the specific.  so in this source I will have a
general 'Application' class which will contain no specific 'radio' code.
the radio code will make use of the general application class.

the program is not yet functional.