1595052: writing strings
I've added the class pywmgeneral.Drawable, offering an interface to pixmaps.

this way a wmdocklib application can have more than one drawable: not only
the main window, but as many pixmaps as are necessary.

the first use of pywmgeneral.Drawable is to hold the graphical
representation of the labels in an application.  Labels are also new to the
library and the intention is that they substitute the putString methods.

Application.addLabel() is used to create a new named label.  the label takes
care of the specified area, or of the area necessary to show the specified
string.

Application.setLabelText() is used to specify a new text for a label.  if
necessary, a new pixmap is created.  the area taken care of by a label
cannot be modified.

during the update cycle, labels that span an area which is too small for
their text are scrolled a pixel at a time.

pywmradio.py uses this functionality.