1723251: it is too difficult to move a button
I've added the two methods addButton and setButtonPattern to the Application class.

the area in the background has to be paintable, not transparent, or the
button will not be displayed nor be sensitive.

patterns are taken from the application and are not stored with the button.

it might be a good idea to introduce an object hierarchy for widgets,
including Labels and Buttons, maybe also flashing buttons, but as far as the
application programmer (of things like the current examples) is concerned,
this is good enough, I'd say.